Moved score panel to its own component.
