+++ src/components/ScoresPanel/index.ts
@@ -1,0 +1,3 @@
+import ScoresPanel from "./ScoresPanel";
+
+export default ScoresPanel;
